Changed num_samples in shader and main to be float

+++ DVR/src/shaders/back_face.frag
@@ -3,7 +3,7 @@
 in vec3 frag_position;
 out vec4 color;
 
-uniform int num_samples;
+uniform float num_samples;
 uniform float smooth_step_start;
 uniform float smooth_step_end;
 
@@ -20,7 +20,7 @@ bool inVolume(vec3 volume_position) {
 void main() {
 
 	vec3 volume_position = texelFetch(front_faces, ivec2(gl_FragCoord), 0).xyz;
-	vec3 step = (frag_position - volume_position) / float(num_samples);
+	vec3 step = (frag_position - volume_position) / num_samples;
 	
 	color = vec4(0.0f);
 	while(inVolume(volume_position)) {
